Finish Lab 3

+++ app/Http/Controllers/PostController.php
@@ -2,21 +2,27 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\CreateNewPost;
 use App\Models\Post;
 use App\Models\User;
+use Auth;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
+    
+
+    use AuthorizesRequests;
     public function index()
     {
-        return view('posts.index', ['posts' => Post::all()]);
+        return view('posts.index', ['posts' => Post::all()->sortByDesc('created_at'), 'userNoPosts' => Auth::user()->posts_count]);
     }
 
     public function create()
     {
-        return view('posts.create', ['users' => User::all()]);
+        return view('posts.create', ['users' => [Auth::user()]]);
     }
 
     public function show(int $id)
@@ -33,9 +39,10 @@ class PostController extends Controller
     {
         try {
             $post = Post::findOrFail($id);
-            return view('posts.edit', ['post' => $post, 'users' => User::all()]);
-        } catch (ModelNotFoundException $err) {
+            $this->authorize('update', $post);
+            return view('posts.edit', ['post' => $post, 'users' => [Auth::user()]]);
 
+        } catch (ModelNotFoundException $err) {
             return abort(404, 'Post not found');
         }
     }
@@ -44,10 +51,11 @@ class PostController extends Controller
         $validatedData = $request->validate([
             'title' => 'required|max:30',
             'body' => 'required|max:255',
-            'user_id' => 'required|numeric'
         ]);
+        $validatedData['user_id'] = Auth::id();
 
-        Post::create($validatedData);
+        $post = Post::create($validatedData);
+        CreateNewPost::dispatch($post);
 
         return response()->redirectTo(route('posts.index'))->with('success', 'Post created successfully.');
 
@@ -55,13 +63,12 @@ class PostController extends Controller
     public function update(Request $request, $id)
     {
 
-
         $post = Post::findOrFail($id);
         $validatedData = $request->validate([
             'title' => 'sometimes|max:30',
             'body' => 'sometimes|max:255',
-            'user_id' => 'sometimes|numeric'
         ]);
+        $validatedData['user_id'] = Auth::id();
 
         $post->update($validatedData);
 
@@ -73,8 +80,9 @@ class PostController extends Controller
     {
         try {
             $post = Post::findOrFail($id);
+            $this->authorize('delete', $post);
+            $post->user()->decrement('posts_count');
             $post->delete();
-            // $res = $res->json(['message' => 'Post deleted successfully.', 'data' => $post]);
 
         } catch (ModelNotFoundException $err) {
             // $res = $res->json(['message' => 'Post not found.'], 404);
